requirements.txt
annotated-types==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
deprecated==1.2.15 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.2.2 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.115.6 ; python_version >= "3.9" and python_version < "4.0"
google-genai==1.11.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
limits==3.14.1 ; python_version >= "3.9" and python_version < "4.0"
lxml==5.3.0 ; python_version >= "3.9" and python_version < "4.0"
motor==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic-settings==2.7.1 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.10.4 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.6.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.9" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
slowapi==0.1.9 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.41.3 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
uvicorn[standard]==0.34.0 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
watchfiles==1.0.3 ; python_version >= "3.9" and python_version < "4.0"
websockets==14.1 ; python_version >= "3.9" and python_version < "4.0"
wrapt==1.17.0 ; python_version >= "3.9" and python_version < "4.0"
